Blood Vendor Medlink Restock (#7331)
# About the pull request

This PR enables the ability for blood vendors to restock via medlinks
(takes 20 minutes from roundstart to begin) and adds med links to most
blood vendor locations on the Almayer. At one point I had it disabled
wanting to encourage more use of the restock carts and possibly for
players to make blood drives in game, but this never became a norm
likely because of the bug that was allowing blood bag reagents to be
refilled by vendors. I will likely explore this again in the future with
the req rework.

# Explain why it's good for the game

At least for the moment, supplies shipside should generally not be
limited. Supply attrition should generally be a result of failing to
transport supplies shipside to groundside.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![image](https://github.com/user-attachments/assets/7f40be38-199f-4a05-951b-fb220fb21e21)

![restock](https://github.com/user-attachments/assets/ee556437-b936-44be-b061-dc70cdde297b)

</details>


# Changelog
:cl: Drathek
qol: Added examine text to vendors that are connected via medlink
balance: Blood vendors restock blood bags when connected via medlink
(requires 20 mins from round start)
maptweak: Added more medlinks to the Almayer
/:cl: