Declutters and lightly reorganizes the HM vendor (#7297)
# About the pull request

Removes a few bloat items and reorganizes the HM vendor, the guiding
philosophy being to take items that are plentiful elsewhere at no cost
out, and to look at the vendor from the perspective of a "new" HM and
try to guide them towards a loadout that will give them a basic level of
effectiveness without railroading them into established meta loadouts.

Notable changes include: 

Gauze and ointment are no longer available for points, due to being
plentiful around the ship and completely inferior-- there is no reason
to take these as an HM.

The G-8A has been removed from the points vendor, this belt is *very*
situational and usually not for beginner loadouts, and is available in
the squad vendors and in req for those who need one, costing 15p is also
highway robbery for what is again, a plentiful free item.

The empty reagent pouch has been removed from the points vendor, you can
get empty reagent pouches from the pouches section of the free vendor,
you can only carry two anyway, and research can print more for unusual
strategies that may require more than 2.

The machete scabbard has been removed from the points vendor, HMs
generally cannot bring a scabbard without heavily compromising their
treatment abilities, and are able to buy a machete pouch which fits
their role better, in addition to these being freely available in high
numbers in req and squad prep.

The fuel tank strap pouch has been removed from the points vendor;
Medics cannot buy flamers in their prep for points, meaning they need to
go to req for one anyway, which has these, it has no reason to be in the
personal vendor for HMs.

The reagent canister pouches have been de-recommended except for Revival
Mix - Tricord; these are still useful, but new players should be
encouraged towards the revival mix.

The 'medical pouch' has been de-recommended, as it is inferior to the
medical kit pouch in every way.

The M276 Lifesaver and Medical Storage Rig are now classed as Mandatory,
other belts do no longer appear in the HM vendor; If you accidentally do
not take your medical belt you are *significantly* disadvantaged and
need to know where to replace these items, which may not be immediately
apparent to the newer players more likely to make this mistake. These
belts are so strong in terms of storage, and because there is almost no
ability that is exclusive to the belt slot, there is no reason not to
take one of these when playing HM except as a meme to prove it's
possible. The other belts are still available in squad prep for
attaching to armor, etc.

I considered removing the spare pill bottles, etc, from the points
vendor since they're available for free elsewhere, but ultimately
decided that at this time I wanted to leave it so that players could get
(almost) all their needed equipment from their vendor-- at a price. (The
lack of IA and defibs in this vendor makes the assertion that it's
currently possible to do so dubious, but those are balance changes I
don't intend to implement without maintainer request.)

I'm open to feedback about these and other suggested changes, if you
have ideas to further improve the HM vendor experience.

# Explain why it's good for the game

The current medical vendor is a wall of unsorted green items that gives
players without an encyclopedic knowledge of what everything does
overload and makes it hard for them to make an effective loadout, since
there aren't refunds or 'do-overs' for item picks, it should be as
difficult as possible to lock yourself out of critical items due to
misinfo.


# Testing Photographs and Procedure

<details>
<summary>Screenshots & Videos</summary>


![image](https://github.com/user-attachments/assets/3560e255-6ad0-41ee-90d3-1404ceeda51e)

</details>


# Changelog
:cl:
qol: the HM vendor's recommendations have been tweaked to be easier to
understand.
/:cl: