Nuclear bomb timer visibility (#7332)
# About the pull request

Gives xenos the same notification about bomb decryption time that preds
and humans get when the decryption is initially activated or activated
after being deactivated.

# Explain why it's good for the game

Parity. Preds get a timer notification before the nuke finishes
decrypting for example. Why do xenos not get this? Also if it's a case
of decryption being reactivated after a long time of deactivation, it's
only fair that the message isn't exactly the same as if they're starting
it for the first time - because the timer can potentially be much
shorter.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
qol: xenos get the same bomb decryption timer notification that preds
get
/:cl: