Fix Splint Removal (#7346)
# About the pull request

Fixes #7291

The check for splints only occurred before the do_after, so if the
splints were removed/destroyed during the do_after, the do_after would
complete and create new splint stacks based on the amount of splints
present before the do_after. Now a check is performed after the do_after
as well to see if the splints still exist.

# Explain why it's good for the game

bugs are bad


# Testing Photographs and Procedure

checked on local removing splints on self/other and destroying splints
during the do_after. Works as intended with removing splints on active
hand as well.

# Changelog

:cl: BOBAMA
fix: Fixes splint removal dupe bug
/:cl:

---------

Co-authored-by: harryob <me@harryob.live>