Synthetic Vendor Update (#7369)
# About the pull request

Adds a pen light to the synthetic points vendor, removes the duplication
of the stethoscope from the clothing vendor (it's a mechanical item now,
not just a clothing item, it should cost points not be essentially
free).

# Explain why it's good for the game

Mechanical items should not be essentially free, also having to bed for
a penlight from medical or take the one from the laundry is silly.

# Testing Photographs and Procedure

Local server did not crash. 


# Changelog

:cl:
add: Pen light added to synthetic points vendor
del: Removed stethoscope from the clothing vendor, left the one in the
points vendor
/:cl: