Adds hallucinations to the memorial slab (#7299)
# About the pull request

**includes code from tgstation, the entire client_image_holder.dm file
is a straight up rip from their codebase with some stuff removed:**
https://github.com/tgstation/tgstation/pull/69706

special thanks to this [ideaguys
thread](https://forum.cm-ss13.com/t/placing-id-tags-on-the-memorial-should-allow-the-ghost-of-that-marine-to-manifest-speak-with-you-like-an-imaginary-friend/10708)
for inspiring me

marines can now inspect (read: click on it when adjacent) the memorial
slab to contemplate some of the names on that list. if they're alone,
ghost images of the dead marines will appear with a random voiceline.
the ghost images will use whatever gear the original mob has on them.

depending on how many marines are listed on the slab, a traumatic
flashback can trigger.
if there are a certain amount of YOUR squad members on the slab, visions
of them will appear around the user and they will get to have one final
goodbye with them as a thank you for not forgetting about them.

# Explain why it's good for the game

it's soulful. more intractability with props is always good, even if
they don't get used that often. it also gives a bit more of a reason to
collect dogtags, so you can see your friends one last time even if
they're gone.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

squad flashback:

outdated, taps no longer plays


https://github.com/user-attachments/assets/35ffc28d-6991-451c-9c5c-82206fadd597





concept:




https://github.com/user-attachments/assets/a1407a28-8cb7-4f76-a924-79ae02ae7383





</details>


# Changelog
:cl:
add: You can now take the time to remember your fallen friends at the
memorial. Don't be alone when trying to remember. It hurts to remember.
/:cl:

---------

Co-authored-by: harryob <me@harryob.live>