CE APOLLO PDA (#7366)
# About the pull request

Adds the same PDA working joes have to CE, but with a unique sprite and
loading screen. A unique tool for CE to interact with AI and working
joes, not to mention CE doesn't have many unique items to begin with,
for a job that has to oversee the AI, having such tool is merely a qol.

Sprites by me.


![sMDwrn0E0K](https://github.com/user-attachments/assets/01ab6372-d57f-41f0-a610-4b177e577d67)


![dreammaker_NLdfLnLh6O](https://github.com/user-attachments/assets/cce0aeeb-2bd3-4e0b-a45e-b57154e9af99)




# Changelog
:cl:
add: adds unique APOLLO pda to CE vendor
/:cl: