Buffs contractor ABR40 (#7286)
# About the pull request

Contractor ABR40 is a bit better, firing faster and doing a bit more
damage, as well as having an integrated magharn, in addition the ABR40
loadout has been rearranged to have more ammo.

# Explain why it's good for the game

ABR40 sucked, as it was balanced mainly as a survivor rifle, so having
this be more equal to marine guns is good, more ammo is important for
its fire-rate as well.

# Changelog
:cl:
balance: The contractor ABR40 now fires slightly faster and does a bit
more damage, also having an integrated magharn.
add: Shuffles around contractor ABR40 loadout, sacrificing the smoke
grenades and holotargeting rounds for a bit more normal ammo.
/:cl:

---------

Co-authored-by: harryob <me@harryob.live>