Updates Tank Primary Camo (#7048)
# About the pull request

* Adds new sprites for LTAA ammo
* Swaps beige camo of primary hardpoint and ammo with the green/grey of
the tank
* Fixes Sprite errors in Tank ammo and primarys

# Explain why it's good for the game

Tank minigun was beige because it was a sprite rip from the painless.
The other tank primarys where made to fit it for some reason and didnt
get a camo fitting to the tank.
Visual consitency is good. 

I know they are still a bit scuffed in one or two places. But this is
just for the camo to be changed. More later.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![grafik](https://github.com/user-attachments/assets/dfb5f552-a8e2-4b94-a063-6aeca324b19f)

![grafik](https://github.com/user-attachments/assets/f7f8b922-01cc-4b98-9e49-c414b1856c5b)

![grafik](https://github.com/user-attachments/assets/219b528a-e427-4988-9171-a7868ceae1e9)

![grafik](https://github.com/user-attachments/assets/1ab52f81-002b-42f2-96de-44d93e2fcb3c)


</details>


# Changelog
:cl:TheManWithNoHands
code: changed the icon path of LTAA mags to be not literally the
painless mag
imageadd: changed LTAA-AP minigun, 20m Autocanon, DRG-N flamer, LTAA-AP
mag, DRG-N mag, to use green tank camo
/:cl: