Gives the Corporate Liaison a map-specific lore-brief about the map. (#7356)
# About the pull request

This PR adds a new variable to map configs, called liaison_briefing,
which is inputted into a piece of paper that spawns on the liaison's fax
machine roundstart.

The idea is to insert a bit of lore about the colony the marines are
going to, as well as to give the liaison some roleplay primers to start
the round with.

If the map does not have a "liaison_briefing" set, it will default to a
placeholder message explaining that there isn't a colony briefing
available for that colony (todo; pretty the placeholder up)

For now, I've only written the trijent one, as it had the most concrete
existing lore I could pull from to write a document about, some of the
other maps (like LV, especially) don't have very well established lore,
and will need working out with the lore guys on what best to include. If
you want to write a briefing for your/your favorite map, feel free, I
recommend preparing it in paperwork sim; [paperwork
sim](http://ps.ss13.net/#instructions), and I'll be more than happy to
convert it to HTML and add it to this PR or PR it for you :)
# Explain why it's good for the game

It always seemed weird to me that a lot of the time the CL has very
little clue about this colony that they supposedly care about and don't
want to get bombed to the stone age, out of everyone on the Almayer, the
Liaison's ties to the company mean he should likely be the best informed
about the colony. To this end, I've created a framework to brief the
liaison about the map and it's interesting quirks, as well as giving
them some fluff about why it may have gone dark(which conveniently never
mentions xenomorphs.) My hope is that this'll aid the liaison in seeming
knowledgable about the colony, as well as provide an avenue to inject
some of the lore directly into the game, instead of being on a somewhat
obscure subpage of a subpage of the wiki.

# Testing Photographs and Procedure

Spawn in on map with a liaison_briefing set, verify the briefing spawns,
and is properly filled out, spawn on a map without a liaison_briefing
set, verify the briefing spawns, and is populated with the placeholder
explaining the lack of a briefing.

<details>
<summary>Screenshots & Videos</summary>
Ingame appearance of the trijent brief:


![zhlFP0BGLC](https://github.com/user-attachments/assets/5e99477a-4b42-4ade-9ab1-55647c8e60da)
paperwork sim view of the trijent brief:


![image](https://github.com/user-attachments/assets/181abb10-cfd1-4c59-8cf4-aea273b9d2e0)
placeholder briefing unavailable brief:


![image](https://github.com/user-attachments/assets/aa90e604-0137-4d85-9697-d1e3a649c65b)

</details>


# Changelog
:cl:detectivegoogle
add: the CL now gets a briefing about the lore of the active map,
assuming one has been written.
/:cl:

---------

Co-authored-by: harryob <me@harryob.live>