Prize pool rebalance for arcade machines (#7150)
# About the pull request
adds a fuck ton of cool and fun prizes to arcade machines that you can
win!
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game

![IMG_2343](https://github.com/user-attachments/assets/f59863c6-d7b9-4a8a-9ded-b84917eb38e7)

I updated the list massively so there is an actual reason to play arcade
machines, with the old prizes there was no incentive as there was no fun
things to win.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
balance: more items have been added to the arcade machine prize pool.
/:cl:

---------

Co-authored-by: BeagleGaming1 <56142455+BeagleGaming1@users.noreply.github.com>