Adds a proper warning dot to thermobaric missile direct fire (#7216)
# About the pull request

Stole code from https://github.com/cmss13-devs/cmss13/pull/5957 to add a
proper warning dot.


# Explain why it's good for the game

Thermobaric is currently the only CAS direct fire missile that cannot be
dodged through the warning intended to make it possible because you
cannot clear the AoE before it pulls you back in. This fixes that.
Essentially the same kinda issue that caused the FATTY rocket to be
removed in the past. No warning = bad.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>



https://github.com/user-attachments/assets/a736acb2-9413-436e-932f-f908d881bee1



</details>


# Changelog
:cl:
add: Added a warning to thermobaric missile direct fire.
/:cl:

---------

Co-authored-by: harryob <me@harryob.live>