Fixes additional projectiles (#7406)
# About the pull request

Fixes Issue #3053 (and potentially #1764)
This PR aims to fix a bug that has been around for a long time which is
explained below.
Changes tested.

The reason why buckshots and any other multi-projectile weapons (or
explosions) are not working correctly is because of the PR that added
random speed variance for projectiles (#1608). This PR caused the
buckshot's additional projectiles to either lag behind or go faster than
the main projectile which stuns the target, causing the additional
projectiles to not hit the target most of the time.
However there is also another factor which causes the additional
projectiles to not hit the target. That is the fact that when the main
projectile hits and stuns the target (makes them fall down to the
floor), the additional projectiles miss the target because they are on
the floor and go over the target.
This causes the shotgun to not deal the intended damage because only the
main projectile hits and barely does any damage.

**This is a bugfix. It doesnt buff the projectile damage in any way. It
technically also only affects stunnable targets because the problem is
targets being stunned and the projectiles missing them.
For example; Hivelords dont get stunned, therefore they dont have an
issue with the additional projectiles.**

# Explain why it's good for the game

A buckshot shouldn't miss a target that you are adjacent to.
The bug causes the buckshot to be completely useless at times, because
it doesnt work as intended.
Even though you are adjacent to your target, the projectiles miss the
target which clearly shouldn't happen.
If you are still confused as to what this PR fixes exactly, videos below
should show the problem.

# Testing Photographs and Procedure

**These are before / after footages. First video is before the fix,
second video is after the fix.**
<details>
<summary>Screenshots & Videos</summary>

Before


https://github.com/user-attachments/assets/364bb19e-25a5-4f6f-9dd3-ce4beea0fcfa

After


https://github.com/user-attachments/assets/b797c3b3-97d2-476b-bac6-f450b976efe6



</details>


# Changelog
:cl: Ansekishoku
fix: Buckshot additional projectiles no more miss the target when they
get stunned by the main projectile.
/:cl:

---------

Co-authored-by: Doubleumc <Doubleumc@users.noreply.github.com>