makes containers more destructible (#7032)
# About the pull request

makes containers destructible by bullets, xeno slashing and adds proper
behavior for explosion act

# Explain why it's good for the game

having containers that can be only removed with acid or instantly
deleted by slightest destruction while invournuable to any other form of
damage is rather silly, especily as it resoults in marines having to use
explosives inside of FOB to get rid of containers in their way.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
balance: containers are destructible with bullets and xeno slashing
/:cl:

---------

Co-authored-by: vincibrv <vojtechvincibr@seznam.cz>
Co-authored-by: BeagleGaming1 <56142455+BeagleGaming1@users.noreply.github.com>