Pre-weeds some areas of Shivas Snowball (#7341)
# About the pull request

Adds some pre-made weeds and a few walls/doors on Shivas Snowball.

# Explain why it's good for the game

Shivas is somewhat disliked by Xenomorph players for a range of reasons,
one of these is the extreme amount of setup time is needed for xenos to
weed the colony plus how easy it is for survivors to move around.

This change will pre-weed the areas that Xenomorphs spawn in, the rough
south and north eastern areas of the map, and add a few set of resin
doors and walls near the main entrances. The idea here is to give the
xenomorphs a small base to start with to make it quicker to start
fortifying hive locations.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
maptweak: The north-east and south-east areas of Shivas Snowball start
of pre-weeded, in addition to a few resin doors and walls near the main
entrances to hive locations.
/:cl:

Co-authored-by: Steelpoint <alexander.henley@hotmai.com>