Tacmap Polish (#7438)
# About the pull request

This PR reworks how the minimap system crops tacmaps to allow maps
larger than the screen. For now the canvas size was increased from 480
to 600 but it is trimmed down to 512. It is also much more strict about
trimming the edge of the map. This allows the Almayer to be fully
displayed in the observer on screen tacmap even though it is larger than
the screenspace by scrolling the map left and right automatically. (I
originally wanted to just scale it down enough to fit on the screen, but
this resulted in inconsistent wall thickness).

The tacmap TGUI is also more polished meaning there is less variation
between each panel and more of the window space is used to accommodate
the maps.

The only sprite change is changing a blank canvas image to be larger.

This also turns out to fix an issue where predators and observers were
using the same tacmap because now the shifting value is part of the hash
making observer maps unique. They cannot share the same map because they
both need unique screen_loc values. I am not completely convinced this
is the only issue that might prevent ghost tacmaps from opening though;
but will need more investigation.

# Explain why it's good for the game

Allows us to more easily support large maps (such as now the Almayer and
its AI Core) so they can still be displayed on the minimap and polishes
the tacmap TGUI panels to be more consistent and utilize space more
efficiently.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![shift](https://github.com/user-attachments/assets/c80805d0-465b-4098-a670-f2b0ce6674e5)

![shift2](https://github.com/user-attachments/assets/03af61c1-1baf-48e3-aa27-0acc9befcd4f)

![minimap](https://github.com/user-attachments/assets/44ca8691-1270-4276-8b0a-7dd225b5526b)

![image](https://github.com/user-attachments/assets/c1e517cc-10a4-42e2-930a-548b0d3e2a1d)

![image](https://github.com/user-attachments/assets/ff225df5-7514-40f1-91cb-2c345819e075)

![image](https://github.com/user-attachments/assets/aafd4fd9-c95c-4bf7-af80-6a7f1df6739f)

![image](https://github.com/user-attachments/assets/523d6d23-fe52-47fe-ad9a-d4ce72a440a5)

![image](https://github.com/user-attachments/assets/cb0716fd-156c-4431-8c00-9b23b89190bc)

![image](https://github.com/user-attachments/assets/5dd614ea-1ae8-420e-8a15-2370bc0ad9a0)

![image](https://github.com/user-attachments/assets/ed98a303-0bf1-42a8-a635-a723ae658ff1)

</details>


# Changelog
:cl: Drathek
ui: The ghost observer tacmap will now scroll both horizontally and
vertically as needed to display a large map.
ui: The TGUI panels for tacmaps and drawings have been polished to
utilize more minimap space fully.
refactor: Refactored minimap generation code to calculate min and max
faster, to more aggressively trim minimaps, and to support larger
minimaps (was 480px now 512px).
fix: Fixes an issue where predator tacmap would take over the ghost
tacmap preventing it from opening for observers
/:cl: