Adds check status and stethoscope usage to examining (#7463)
# About the pull request

human mobs can now directly use check status/the stethoscope attached to
their uniform if they examine a mob

# Explain why it's good for the game

stethoscope is pretty good but having to fiddle around with your uniform
to take it off constantly is annoying. check status is a pretty
informational verb but the problem is that it's a verb, so it's stuck
behind a right click menu. this is more direct and lets people use them
a tad more easily

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>



https://github.com/user-attachments/assets/17b99ed0-38d3-4c0f-8913-bfde24af82f0



</details>


# Changelog
:cl:
qol: You can now use the 'check status' verb and any stethoscope you
might have attached to your uniform directly by examining an adjacent
human.
/:cl: